Initial Attempt at using the Streaming API
I create a streaming_update.py file which borrows from the update.py. It listens to the streaming API for stars, and posts. All that it does right now is find posts that aren't replies and sends them to process_post.

I also added manage.py. This makes it easier to run one off commands in the context of flask. This will require that you update your environment because I added a flask extension to the requirements.